requirements.txt
psycopg2==2.6.2
django-councilmatic==0.9.3
django-councilmatic-notifications<0.2
pysolr===3.6.0
django-haystack==2.6.1
raven==5.30.0
google-api-python-client==1.6.4
icalendar==4.0.0
gunicorn==19.6.0
pytest==3.2.5
pytest-django==3.1.2
pytest-mock==1.6.3
